Web App - Nepali Digits/requirements.txt
Flask
Keras
numpy
tensorflow
Pillow
h5py
opencv-python